requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.0
langchain-huggingface==0.0.3
chromadb==0.4.22
sentence-transformers==2.2.2
tavily-python==0.5.0
arxiv==2.1.0
pypdf==3.17.4
python-dotenv==1.0.0
openai==1.30.0